requirements.txt
absl-py==1.3.0
ahk==0.12.0
aiohttp==3.8.3
aiosignal==1.3.1
ansi2html==1.8.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.2.3
asn1crypto==1.5.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==3.3.0
bs4==0.0.1
cachetools==5.2.0
careerjet-api==3.2
cattrs==1.5.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cloud-sql-python-connector==1.1.0
cloudpickle==2.2.0
colorama==0.4.4
comtypes==1.1.7
cryptography==39.0.0
dash==2.7.1
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
finnhub-python==2.4.15
Flask==2.2.2
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==3.4.2
google-cloud-core==2.3.2
google-cloud-secret-manager==2.15.0
google-cloud-talent==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
h5py==3.7.0
httplib2==0.21.0
idna==2.10
importlib-metadata==5.0.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.1
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.0
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter-server==1.23.5
jupyter_client==7.4.1
jupyter_core==5.1.5
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupyterlab_server==2.19.0
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mistune==2.0.4
mlagents==0.28.0
mlagents-envs==0.28.0
multidict==6.0.4
nano==0.10.0
nbclassic==0.5.1
nbclient==0.5.3
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.4.0
notebook_shim==0.2.2
numba==0.56.4
numpy==1.20.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.3
pg8000==1.29.4
photoshop-python-api==0.16.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.6.2
plotly==5.13.0
plotly-express==0.4.1
pooch==1.6.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.25.1
Pygments==2.9.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytube==12.1.2
pytz==2022.7.1
pywin32==301
pywinauto==0.6.8
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.1.0
QtPy==1.9.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.6.3
scramp==1.4.4
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
SQLAlchemy==1.4.46
statsmodels==0.13.5
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.7.1+cu110
tornado==6.2
traitlets==5.8.1
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.5
voila==0.4.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==3.5.1
yarl==1.8.2
youtube2text==0.0.9
zipp==3.10.0
